Trying to figure out why location editing is now broken

+++ Projektikansio/MyApi/Controllers/MatkakohdesController.cs
@@ -74,7 +74,7 @@ namespace MyApi.Controllers
             return matkakohde.toMatkakohdeDTO();
         }
 
-        // PUT: api/Matkakohdes/5 vai ...user/id?
+        // PUT: api/Matkakohdes/5 
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<ActionResult<matkakohdeDTO>> PutLocation(long id, matkakohdeDTO m)
@@ -83,8 +83,8 @@ namespace MyApi.Controllers
             if (location == null) return NotFound();
 
             //Sellaista matkakohdetta, johon liittyy joku matkakertomus, ei saa poistaa tai päivittää
-           // Tarina? t = await _context.Tarinas.Where(a => a.Idmatkakohde == id).FirstOrDefaultAsync();
-           // if(t != null) { return Ok(); } 
+            Tarina? t = await _context.Tarinas.Where(a => a.Idmatkakohde == id).FirstOrDefaultAsync();
+            if(t != null) { return Ok(); } 
 
             if (m == null) return BadRequest(); 
             else
@@ -146,8 +146,8 @@ namespace MyApi.Controllers
             if (m == null) return NotFound();
 
             //Sellaista matkakohdetta, johon liittyy joku matkakertomus, ei saa poistaa tai päivittää
-          //  Tarina? t = await _context.Tarinas.Where(a => a.Idmatkakohde == id).FirstOrDefaultAsync();
-           // if (t != null) { return BadRequest(); }
+            Tarina? t = await _context.Tarinas.Where(a => a.Idmatkakohde == id).FirstOrDefaultAsync();
+            if (t != null) { return BadRequest(); }
 
             _context.Matkakohdes.Remove(m);
             await _context.SaveChangesAsync();
